# specs/001-volunteer-care-report/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StrayHub CRM Care Report API
  version: 0.1.0
  description: >-
    志工日常照護回報與動物近期歷程的 CRM 邊界契約。一般收容所操作由
    FastAPI 重新驗證 Session、Membership、角色與 Active Shelter Context；
    PLATFORM_ADMIN 使用平台級 PLATFORM Scope，且所有跨收容所操作都留下 Audit Record。
...
servers:
  - url: http://localhost:8000
    description: 本機開發
...
security:
  - bearerAuth: []
paths:
...
  /v1/management/animals/{animalId}/medical-records:
    get:
      tags: [Medical Records]
      operationId: listMedicalRecords
      parameters:
        - $ref: '#/components/parameters/AnimalId'
        - { name: occurred_from, in: query, schema: { type: string, format: date } }
        - { name: occurred_to, in: query, schema: { type: string, format: date } }
        - { name: record_type, in: query, schema: { $ref: '#/components/schemas/MedicalRecordType' } }
        - { name: search, in: query, schema: { type: string } }
        - { name: include_archived, in: query, schema: { type: boolean, default: false } }
      responses:
        '200': { description: 醫療歷史, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecordPage' } } } }
    post:
      tags: [Medical Records]
      operationId: createMedicalRecord
      parameters: [{ $ref: '#/components/parameters/AnimalId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecordCreate' } } } }
      responses:
        '201': { description: 建立醫療歷史, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecord' } } } }
  /v1/management/medical-records/{recordId}:
    get:
      tags: [Medical Records]
      operationId: getMedicalRecord
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      responses: { '200': { description: 醫療紀錄, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecord' } } } } }
    patch:
      tags: [Medical Records]
      operationId: updateMedicalRecord
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecordUpdate' } } } }
      responses: { '200': { description: 更新醫療紀錄, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecord' } } } }, '409': { $ref: '#/components/responses/Conflict' } }
  /v1/management/medical-records/{recordId}/archive:
    post:
      tags: [Medical Records]
      operationId: archiveMedicalRecord
      parameters: [{ $ref: '#/components/parameters/RecordId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecordArchive' } } } }
      responses: { '200': { description: 封存醫療紀錄, content: { application/json: { schema: { $ref: '#/components/schemas/MedicalRecord' } } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 簽署的短效 Access Token；每次受保護 Request 仍必須通過 Server-side Session 與 Organization Scope 驗證。
  parameters:
...
    AnimalId:
      {
        name: animalId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
    RecordId: { name: recordId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Conflict:
      description: Idempotency、stale version、operation payload 或資料狀態衝突
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        operation_id: { type: [string, "null"], format: uuid }
        details: { type: object, additionalProperties: true }
...
    MedicalRecordType:
      type: string
      enum: [visit, medication, vaccination, examination, weight, surgery, other]
    MedicalRecord:
      type: object
      required: [id, animal_id, occurred_at, occurred_timezone, record_type, title, content, status, version, created_by_user_id, created_at, updated_by_user_id, updated_at, media_ids]
      properties:
        id: { type: string, format: uuid }
        animal_id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time }
        occurred_timezone: { type: string }
        record_type: { $ref: '#/components/schemas/MedicalRecordType' }
        title: { type: string }
        content: { type: string }
        clinic: { type: [string, 'null'] }
        veterinarian: { type: [string, 'null'] }
        weight_kg: { type: [number, 'null'], exclusiveMinimum: 0 }
        status: { type: string, enum: [active, archived] }
        version: { type: integer, minimum: 1 }
        created_by_user_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_by_user_id: { type: string, format: uuid }
        updated_at: { type: string, format: date-time }
        archived_at: { type: [string, 'null'], format: date-time }
        archive_reason: { type: [string, 'null'] }
        media_ids: { type: array, items: { type: string, format: uuid } }
    MedicalRecordCreate:
      type: object
      required: [occurred_at, record_type, title, content]
      additionalProperties: false
      properties:
        occurred_at: { type: string, format: date-time }
        record_type: { $ref: '#/components/schemas/MedicalRecordType' }
        title: { type: string, minLength: 1, maxLength: 200 }
        content: { type: string, minLength: 1, maxLength: 10000 }
        clinic: { type: [string, 'null'] }
        veterinarian: { type: [string, 'null'] }
        weight_kg: { type: [number, 'null'], exclusiveMinimum: 0 }
        media_ids: { type: array, maxItems: 8, items: { type: string, format: uuid } }
    MedicalRecordUpdate:
      allOf:
        - $ref: '#/components/schemas/MedicalRecordCreate'
        - type: object
          required: [expected_version, reason]
          properties:
            expected_version: { type: integer, minimum: 1 }
            reason: { type: string, minLength: 1 }
    MedicalRecordArchive:
      type: object
      required: [expected_version, reason]
      properties:
        expected_version: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 1 }
    MedicalRecordPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/MedicalRecord' } }
        next_cursor: { type: [string, 'null'] }
...